Fix some compiler warnings.

+++ src/auth.h
@@ -4,10 +4,11 @@
 #pragma once
 
 #include <stdbool.h>
+#include <stdint.h>
 
 extern int  netrc_lookup (const char *host, const char **user, const char **passw);
-extern int  authinfo_lookup (const char *host, const char **user, const char **passw, int *port);
-extern int  envtool_cfg_lookup (const char *host, const char **user, const char **passw, int *port);
+extern int  authinfo_lookup (const char *host, const char **user, const char **passw, uint16_t *port);
+extern int  envtool_cfg_lookup (const char *host, const char **user, const char **passw, uint16_t *port);
 
 extern void netrc_exit (void);
 extern void authinfo_exit (void);
